Seperate IO files into impl. and header files
